encode texcoord to half because repeat_wrap manually or by GLM do not appear to be correct.

+++ JustRayToolKit/Vertex.h
@@ -1,5 +1,6 @@
 #pragma once
 #include "MathUtil.h"
+#include "Half\half.h"
 #include <memory>
 namespace JustRay
 {
@@ -9,8 +10,8 @@ struct Vertex
     Int_2_10_10_10 normal;
     Int_2_10_10_10 tangent;
     Int_2_10_10_10 bitangent;
-    unsigned short tx;
-    unsigned short ty;
+    half tx;
+    half ty;
     bool operator<(const Vertex& rhs) const
     {
         return std::memcmp((void*)this, (void*)&rhs, sizeof(Vertex)) < 0;
